// apps/api/src/modules/openai/dto/create-response-dtos/ApplyPatchCallDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsIn, IsOptional, IsString } from 'class-validator';

import { CreateFileDto } from './CreateFileDto';
import { DeleteFileDto } from './DeleteFileDto';
import { UpdateFileDto } from './UpdateFileDto';

export class ApplyPatchCallDto {
  /** The unique ID of the apply patch tool call generated by the model. */
  @ApiProperty({ description: `The unique ID of the apply patch tool call generated by the model.`, type: 'string' })
  @IsString()
  call_id!: string;

  /** The specific create, delete, or update instruction for the apply_patch tool
call. */
  @ApiProperty({ description: `The specific create, delete, or update instruction for the apply_patch tool
call.` })
  operation!: CreateFileDto | DeleteFileDto | UpdateFileDto;

  /** The status of the apply patch tool call. One of `in_progress` or `completed`. */
  @ApiProperty({ description: `The status of the apply patch tool call. One of \`in_progress\` or \`completed\`.`, enum: ['in_progress', 'completed'] })
  @IsIn(['in_progress', 'completed'])
  status!: 'in_progress' | 'completed';

  /** The type of the item. Always `apply_patch_call`. */
  @ApiProperty({ description: `The type of the item. Always \`apply_patch_call\`.`, example: 'apply_patch_call' })
  @Equals('apply_patch_call')
  type!: 'apply_patch_call';

  /** The unique ID of the apply patch tool call. Populated when this item is returned
via API. */
  @ApiProperty({ required: false, description: `The unique ID of the apply patch tool call. Populated when this item is returned
via API.` })
  @IsOptional()
  id?: null | string;
}