// apps/api/src/modules/openai/dto/create-response-dtos/FunctionCallOutputDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsOptional, IsString } from 'class-validator';

import { ResponseInputTextContentDto } from './ResponseInputTextContentDto';
import { ResponseInputImageContentDto } from './ResponseInputImageContentDto';
import { ResponseInputFileContentDto } from './ResponseInputFileContentDto';

export class FunctionCallOutputDto {
  /** The unique ID of the function tool call generated by the model. */
  @ApiProperty({ description: `The unique ID of the function tool call generated by the model.`, type: 'string' })
  @IsString()
  call_id!: string;

  /** Text, image, or file output of the function tool call. */
  @ApiProperty({ description: `Text, image, or file output of the function tool call.` })
  output!: string | ResponseInputTextContentDto | ResponseInputImageContentDto | ResponseInputFileContentDto[];

  /** The type of the function tool call output. Always `function_call_output`. */
  @ApiProperty({ description: `The type of the function tool call output. Always \`function_call_output\`.`, example: 'function_call_output' })
  @Equals('function_call_output')
  type!: 'function_call_output';

  /** The unique ID of the function tool call output. Populated when this item is
returned via API. */
  @ApiProperty({ required: false, description: `The unique ID of the function tool call output. Populated when this item is
returned via API.` })
  @IsOptional()
  id?: null | string;

  /** The status of the item. One of `in_progress`, `completed`, or `incomplete`.
Populated when items are returned via API. */
  @ApiProperty({ required: false, description: `The status of the item. One of \`in_progress\`, \`completed\`, or \`incomplete\`.
Populated when items are returned via API.` })
  @IsOptional()
  status?: null | 'in_progress' | 'completed' | 'incomplete';
}