// apps/api/src/modules/openai/dto/create-response-dtos/LocalShellCallDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsIn, IsString, ValidateNested } from 'class-validator';
import { Type } from 'class-transformer';

import { ActionDto } from './ActionDto';

export class LocalShellCallDto {
  /** The unique ID of the local shell call. */
  @ApiProperty({ description: `The unique ID of the local shell call.`, type: 'string' })
  @IsString()
  id!: string;

  /** Execute a shell command on the server. */
  @ApiProperty({ description: `Execute a shell command on the server.`, type: () => ActionDto })
  @ValidateNested()
  @Type(() => ActionDto)
  action!: ActionDto;

  /** The unique ID of the local shell tool call generated by the model. */
  @ApiProperty({ description: `The unique ID of the local shell tool call generated by the model.`, type: 'string' })
  @IsString()
  call_id!: string;

  /** The status of the local shell call. */
  @ApiProperty({ description: `The status of the local shell call.`, enum: ['in_progress', 'completed', 'incomplete'] })
  @IsIn(['in_progress', 'completed', 'incomplete'])
  status!: 'in_progress' | 'completed' | 'incomplete';

  /** The type of the local shell call. Always `local_shell_call`. */
  @ApiProperty({ description: `The type of the local shell call. Always \`local_shell_call\`.`, example: 'local_shell_call' })
  @Equals('local_shell_call')
  type!: 'local_shell_call';
}